Refactor AssigneeSelect component and import dependencies
This commit refactors the AssigneeSelect component in the app/issues/[id]/AssigneeSelect.tsx file. It imports the User type from the @prisma/client module and the axios, useEffect, and useState functions from their respective modules. It also adds a fetchUsers function that makes a GET request to the /api/users endpoint and sets the fetched data to the users state. The Select.Item component is updated to dynamically render the users' names. Additionally, the ASsigneeSelect component is renamed to AssigneeSelect.